cleanup no longer relevant experiment files and examples. move `jsx_transform4.ts` to under `/src/jsx/` as `hyperscript.ts`.
