--- examples/1/index.html
@@ -1,25 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-	<meta charset="utf-8" />
-	<meta name="viewport" content="width=device-width, initial-scale=1.0" />
-	<title>example 1: seconds counter</title>
-	<!-- <script src="./index.ts" type="module"></script> -->
-	<script>
-		globalThis.addEventListener("DOMContentLoaded", async () => {
-			const { MyComponent, AnotherComponent } = await import("./index.tsx")
-
-			document.body.appendChild(MyComponent())
-			document.body.appendChild(AnotherComponent)
-			Object.assign(globalThis, {})
-		})
-	</script>
-
-	<style></style>
-</head>
-
-<body style="width: 100vw; height: 100vh;">
-</body>
-
-</html>
